Some stuff

+++ src/kernel/idt.h
@@ -2,5 +2,5 @@
 
 #include "types.h"
 
-void idt_gate_set(int n, uint32_t handler);
+void idt_gate_set(int n, uintptr_t handler);
 void idt_install();
